refactor: rename SummaryPage to Summary
